I suggest we use this PR to keep track of the progress on implementing the search for identifiable functions of parameters. @pogudingleb @ChrisRackauckas 

As a general outline, three pieces of code are needed: the GB backend in [Groebner.jl](https://github.com/sumiya11/Groebner.jl), the evaluation-interpolation scheme in [ParamPunPam.jl](https://github.com/sumiya11/ParamPunPam.jl), and the frontend in [StructuralIdentifiability.jl](https://github.com/SciML/StructuralIdentifiability.jl). StructuralIdentifiability.jl will use ParamPunPam.jl to compute Groebner bases with parameters, which, in turn, will use Groebner.jl as a building block. Both ParamPunPam.jl and Groebner.jl are already usable for this purpose to some extent. 

I think a good goal for the very nearest future is to have a working prototype in SI.jl and to look at the challenging examples more closely to decide what needs to be optimized. I suggest the first milestones:

- [ ] Implementing a simple interface in StructuralIdentifiability.jl #187
- [x] Finishing the prep work in Groebner.jl, https://github.com/sumiya11/Groebner.jl/pull/67
- [x] Implementing tracing in Groebner.jl

-- Alex